read description
Fixed typos so error messages pop up for security question and answer
when they are left blank. However, I can't get them to display in red
like the error message for username. Also, I can't get the box for
security question and answer to be as wide as the box for username.